+++ cms.ts
@@ -1,5 +1,5 @@
-export { default as adapter } from "https://cdn.jsdelivr.net/gh/lumeland/cms@v0.3.1/adapters/lume.ts";
-export { default } from "https://cdn.jsdelivr.net/gh/lumeland/cms@v0.3.1/mod.ts";
-export { default as Kv } from "https://cdn.jsdelivr.net/gh/lumeland/cms@v0.3.1/storage/kv.ts";
-export { default as Fs } from "https://cdn.jsdelivr.net/gh/lumeland/cms@v0.3.1/storage/fs.ts";
-export { default as GitHub } from "https://cdn.jsdelivr.net/gh/lumeland/cms@v0.3.1/storage/github.ts";
+export { default as adapter } from "https://cdn.jsdelivr.net/gh/lumeland/cms@v0.3.4/adapters/lume.ts";
+export { default } from "https://cdn.jsdelivr.net/gh/lumeland/cms@v0.3.4/mod.ts";
+export { default as Kv } from "https://cdn.jsdelivr.net/gh/lumeland/cms@v0.3.4/storage/kv.ts";
+export { default as Fs } from "https://cdn.jsdelivr.net/gh/lumeland/cms@v0.3.4/storage/fs.ts";
+export { default as GitHub } from "https://cdn.jsdelivr.net/gh/lumeland/cms@v0.3.4/storage/github.ts";