+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ## [2.1.2] - Unreleased
 ### Fixed
 - UnoCSS plugin: the async process of adding `<style>` elements is not awaited [#578].
+- Updated dependencies: `std`, `terser`, `cms`, `postcss`, `react` types, `unocss`.
 
 ## [2.1.1] - 2024-03-01
 ### Fixed