+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/).
 Go to the `v1` branch to see the changelog of Lume 1.
 
 ## [2.1.2] - Unreleased
+### Added
+- `transform_images` plugin: added the `.webp` extension to the default options.
+
 ### Fixed
 - UnoCSS plugin: the async process of adding `<style>` elements is not awaited [#578].
 - Updated dependencies: `std`, `terser`, `cms`, `postcss`, `react` types, `unocss`, `vento`.