+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ### Fixed
 - UnoCSS plugin: the async process of adding `<style>` elements is not awaited [#578].
 - Updated dependencies: `std`, `terser`, `cms`, `postcss`, `react` types, `unocss`, `vento`.
+- Replace Sharp with `svg2png` as the library to convert svg to png.
 
 ## [2.1.1] - 2024-03-01
 ### Fixed