+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - UnoCSS plugin: the async process of adding `<style>` elements is not awaited [#578].
 - Updated dependencies: `std`, `terser`, `cms`, `postcss`, `react` types, `unocss`, `vento`, `date-fns`, `sass`, `terser`.
 - Replace Sharp with `svg2png` as the library to convert svg to png.
+- Init script creates always the `_cms.ts` file.
 
 ## [2.1.1] - 2024-03-01
 ### Fixed